sleep moved in the utility class + EventsActivityTest wip
