+++ app/src/androidTest/java/com/sdpteam/connectout/authentication/ProfileGreetingsActivityTest.java
@@ -5,7 +5,7 @@ import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
-import static com.sdpteam.connectout.utils.FutureUtil.fJoin;
+import static com.sdpteam.connectout.utils.FutureUtils.fJoin;
 import static org.junit.Assert.assertEquals;
 
 import java.util.concurrent.CompletableFuture;
@@ -28,7 +28,6 @@ public class ProfileGreetingsActivityTest {
     @Rule
     public ActivityScenarioRule<ProfileGreetingActivity> activityScenarioRule =
             new ActivityScenarioRule<>(ProfileGreetingActivity.class);
-
 
     public void testGreetingActivity() {
         Intent intent = new Intent(ApplicationProvider.getApplicationContext(), ProfileGreetingActivity.class);
