+++ app/src/androidTest/java/com/sdpteam/connectout/chat/ChatActivityTest.java
@@ -8,6 +8,7 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
+import static com.sdpteam.connectout.utils.FutureUtils.waitABit;
 import static org.hamcrest.Matchers.anything;
 
 import org.junit.After;
@@ -19,7 +20,6 @@ import org.junit.runner.RunWith;
 import com.sdpteam.connectout.R;
 
 import android.content.Intent;
-import android.os.SystemClock;
 import androidx.test.core.app.ApplicationProvider;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.matcher.ViewMatchers;
@@ -38,14 +38,14 @@ public class ChatActivityTest {
     public final void setUp() {
         Intents.init();
         model.emptyTestMode();
-        SystemClock.sleep(1000);
+        waitABit();
     }
 
     @After
     public final void tearDown() {
         Intents.release();
         model.emptyTestMode();
-        SystemClock.sleep(1000);
+        waitABit();
     }
 
     @Test
@@ -56,7 +56,7 @@ public class ChatActivityTest {
         model.saveMessage(new ChatMessage("Dylan", "Dylan userId", "Hey I m Dylan", nameOfTestChat));
         model.saveMessage(new ChatMessage("Alice", "Alice userId", "I think we should go swimming one of these days", nameOfTestChat));
         model.saveMessage(new ChatMessage("Dylan", "Dylan userId", "Excellent Idea", nameOfTestChat));
-        SystemClock.sleep(1000);
+        waitABit();
         checkListViewValue(0, "Hey I m Bob", "Bob");
         checkListViewValue(1, "Hey I m Alice", "Alice");
         checkListViewValue(2, "Hey I m Dylan", "Dylan");
@@ -65,12 +65,12 @@ public class ChatActivityTest {
 
         onView(withId(R.id.chat_input)).perform(typeText("Excellent Idea"));
         onView(withId(R.id.chat_fab)).perform(click());
-        SystemClock.sleep(1000);
+        waitABit();
         checkListViewValue(5, "Excellent Idea", "You");
 
         onView(withId(R.id.chat_input)).perform(typeText("Hi beautiful people"));
         onView(withId(R.id.chat_fab)).perform(click());
-        SystemClock.sleep(1000);
+        waitABit();
         checkListViewValue(6, "Hi beautiful people", "You");
     }
