+++ app/src/androidTest/java/com/sdpteam/connectout/event/EventsActivityTest.java
@@ -6,33 +6,17 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.intent.Intents.intended;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasComponent;
 import static androidx.test.espresso.intent.matcher.IntentMatchers.hasExtra;
-import static androidx.test.espresso.matcher.ViewMatchers.assertThat;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static com.sdpteam.connectout.event.viewer.EventActivity.PASSED_ID_KEY;
-import static com.sdpteam.connectout.utils.FutureUtil.fJoin;
+import static com.sdpteam.connectout.utils.FutureUtils.fJoin;
+import static com.sdpteam.connectout.utils.FutureUtils.waitABit;
 import static com.sdpteam.connectout.utils.WithIndexMatcher.withIndex;
 import static org.hamcrest.Matchers.equalTo;
 
-import android.os.Handler;
-import android.os.Looper;
-
-import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
-import androidx.test.espresso.intent.Intents;
-import androidx.test.espresso.matcher.ViewMatchers;
-import androidx.test.ext.junit.rules.ActivityScenarioRule;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-import com.sdpteam.connectout.R;
-import com.sdpteam.connectout.event.nearbyEvents.EventsActivity;
-import com.sdpteam.connectout.event.nearbyEvents.EventsViewModel;
-import com.sdpteam.connectout.event.nearbyEvents.filter.EventFilter;
-import com.sdpteam.connectout.event.nearbyEvents.filter.ProfilesFilter;
-import com.sdpteam.connectout.event.viewer.EventActivity;
-import com.sdpteam.connectout.profile.ProfileActivity;
-import com.sdpteam.connectout.utils.Chronometer;
-import com.sdpteam.connectout.utils.LiveDataTestUtil;
+import java.util.List;
+import java.util.Random;
 
 import org.hamcrest.Matchers;
 import org.junit.After;
@@ -41,24 +25,43 @@ import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
-import java.util.List;
-import java.util.concurrent.TimeoutException;
+import com.sdpteam.connectout.R;
+import com.sdpteam.connectout.event.nearbyEvents.EventsActivity;
+import com.sdpteam.connectout.event.nearbyEvents.EventsViewModel;
+import com.sdpteam.connectout.event.nearbyEvents.filter.EventFilter;
+import com.sdpteam.connectout.event.nearbyEvents.filter.ProfilesFilter;
+import com.sdpteam.connectout.event.viewer.EventActivity;
+import com.sdpteam.connectout.profile.ProfileActivity;
+import com.sdpteam.connectout.utils.LiveDataTestUtil;
+
+import android.os.Handler;
+import android.os.Looper;
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule;
+import androidx.test.espresso.intent.Intents;
+import androidx.test.espresso.matcher.ViewMatchers;
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 @RunWith(AndroidJUnit4.class)
 public class EventsActivityTest {
 
-    @Rule public InstantTaskExecutorRule instantTaskExecutorRule = new InstantTaskExecutorRule();
+    private final EventRepository eventDs = new EventFirebaseDataSource();
+    @Rule
+    public InstantTaskExecutorRule instantTaskExecutorRule = new InstantTaskExecutorRule();
     @Rule
     public ActivityScenarioRule<EventsActivity> testRule = new ActivityScenarioRule<>(EventsActivity.class);
 
     @Before
     public void setup() {
         Intents.init();
+        eventDs.saveEvent(new Event("tmp_event_id_" + new Random().nextInt(), "Title", "Desc", null, "aa"));
+        waitABit();
     }
 
     @After
     public void cleanup() {
         Intents.release();
+        // TODO delete this tmp event
     }
 
     @Test
@@ -111,28 +114,19 @@ public class EventsActivityTest {
     }
 
     @Test
-    public void clickingOnAEventLaunchesRightEventPage() throws TimeoutException {
+    public void clickingOnAEventLaunchesRightEventPage() {
         onView(withId(R.id.list_switch_button)).perform(click());
         onView(ViewMatchers.withId(R.id.events_list_view)).check(matches(isDisplayed()));
         onView(withId(R.id.events_list_view)).check(matches(isDisplayed()));
 
-        EventsViewModel model = new EventsViewModel(new EventFirebaseDataSource());
+        EventsViewModel model = new EventsViewModel(eventDs);
+
         Handler handler = new Handler(Looper.getMainLooper());
         handler.post(() -> model.refreshEvents()); // set up the live data in main thread (because we cannot invoke [LiveData].setValue on a background thread)
-        List<Event> list;
-        Chronometer chronometer = new Chronometer();
-        chronometer.start();
-        chronometer.setThreshold(10000);
-        do {
-             list = LiveDataTestUtil.getOrAwaitValue(model.getEventListLiveData());
-        }while (list.size() == 0 && !chronometer.hasExceededThreshold());
-        chronometer.stop();
-        if(chronometer.hasExceededThreshold()){
-            throw new TimeoutException();
-        }
+        waitABit();
+        List<Event> list = model.getEventListLiveData().getValue();
 
         int userIndexToCheck = 0;
-
         String expectedUid = list.get(userIndexToCheck).getId();
 
         onView(withIndex(withId(R.id.event_list_event_title), userIndexToCheck)).perform(click());
