+++ app/src/androidTest/java/com/sdpteam/connectout/event/filter/EventLocationFilterTest.java
@@ -3,11 +3,11 @@ package com.sdpteam.connectout.event.filter;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
+import org.junit.Test;
+
 import com.sdpteam.connectout.event.Event;
 import com.sdpteam.connectout.event.nearbyEvents.filter.EventLocationFilter;
 import com.sdpteam.connectout.event.nearbyEvents.map.GPSCoordinates;
-
-import org.junit.Test;
 
 public class EventLocationFilterTest {
