+++ app/src/androidTest/java/com/sdpteam/connectout/event/filter/ProfilesNameFilterTest.java
@@ -3,17 +3,24 @@ package com.sdpteam.connectout.event.filter;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import com.sdpteam.connectout.event.nearbyEvents.filter.ProfilesNameFilter;
-import com.sdpteam.connectout.profile.Profile;
-
-import org.junit.Test;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import org.junit.Test;
+
+import com.sdpteam.connectout.event.nearbyEvents.filter.ProfilesNameFilter;
+import com.sdpteam.connectout.profile.Profile;
 
 public class ProfilesNameFilterTest {
+
+    private static List<Profile> createListWithMatchingName() {
+        return Arrays.asList(
+                new Profile("John", "John", "mail", "bio", Profile.Gender.MALE, 4.5, 10, ""),
+                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 3.0, 10, ""),
+                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 3.8, 10, "")
+        );
+    }
 
     @Test
     public void testMatchingName() {
@@ -31,13 +38,5 @@ public class ProfilesNameFilterTest {
     public void testEmptyProfileList() {
         ProfilesNameFilter filter = new ProfilesNameFilter("John");
         assertFalse(filter.test(new ArrayList<>()));
-    }
-
-    private static List<Profile> createListWithMatchingName() {
-        return Arrays.asList(
-                new Profile("John", "John", "mail", "bio", Profile.Gender.MALE, 4.5, 10, ""),
-                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 3.0, 10, ""),
-                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 3.8, 10, "")
-        );
     }
 }
