+++ app/src/androidTest/java/com/sdpteam/connectout/event/filter/ProfilesRatingFilterTest.java
@@ -3,17 +3,40 @@ package com.sdpteam.connectout.event.filter;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import com.sdpteam.connectout.event.nearbyEvents.filter.ProfilesRatingFilter;
-import com.sdpteam.connectout.profile.Profile;
-
-import org.junit.Test;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import org.junit.Test;
+
+import com.sdpteam.connectout.event.nearbyEvents.filter.ProfilesRatingFilter;
+import com.sdpteam.connectout.profile.Profile;
 
 public class ProfilesRatingFilterTest {
+
+    private static List<Profile> createGreaterList() {
+        return Arrays.asList(
+                new Profile("John", "Doe", "mail", "bio", Profile.Gender.MALE, 4.5, 10, ""),
+                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 3.0, 10, ""),
+                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 3.8, 10, "")
+        );
+    }
+
+    private static List<Profile> createMixList() {
+        return Arrays.asList(
+                new Profile("John", "Doe", "mail", "bio", Profile.Gender.MALE, 4.5, 10, ""),
+                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 2.0, 10, ""),
+                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 2.76, 10, "")
+        );
+    }
+
+    private static List<Profile> createSmallerList() {
+        return Arrays.asList(
+                new Profile("John", "Doe", "mail", "bio", Profile.Gender.MALE, 1.5, 10, ""),
+                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 2.0, 10, ""),
+                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 2.76, 10, "")
+        );
+    }
 
     @Test
     public void testAllProfilesWithRatingEqualOrGreater() {
@@ -43,31 +66,5 @@ public class ProfilesRatingFilterTest {
     public void testNullValue() {
         ProfilesRatingFilter filter = new ProfilesRatingFilter(null);
         assertTrue(filter.test(createGreaterList()));
-    }
-
-    private static List<Profile> createGreaterList() {
-        return Arrays.asList(
-                new Profile("John", "Doe", "mail", "bio", Profile.Gender.MALE, 4.5, 10, ""),
-                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 3.0, 10, ""),
-                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 3.8, 10, "")
-        );
-    }
-
-    private static List<Profile> createMixList() {
-        return Arrays.asList(
-                new Profile("John", "Doe", "mail", "bio", Profile.Gender.MALE, 4.5, 10, ""),
-                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 2.0, 10, ""),
-                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 2.76, 10, "")
-        );
-
-    }
-
-    private static List<Profile> createSmallerList() {
-        return Arrays.asList(
-                new Profile("John", "Doe", "mail", "bio", Profile.Gender.MALE, 1.5, 10, ""),
-                new Profile("Jane", "Doe", "mail", "bio", Profile.Gender.FEMALE, 2.0, 10, ""),
-                new Profile("Bob", "Doe", "mail", "bio", Profile.Gender.OTHER, 2.76, 10, "")
-        );
-
     }
 }
