+++ app/src/androidTest/java/com/sdpteam/connectout/map/GPSCoordinatesTest.java
@@ -6,10 +6,10 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertTrue;
 
+import org.junit.Test;
+
 import com.google.android.gms.maps.model.LatLng;
 import com.sdpteam.connectout.event.nearbyEvents.map.GPSCoordinates;
-
-import org.junit.Test;
 
 public class GPSCoordinatesTest {
     private final static LatLng VALID_POSITION = new LatLng(37.7749, -122.4194);
