+++ app/src/androidTest/java/com/sdpteam/connectout/map/MapFragmentTest.java
@@ -5,6 +5,7 @@ import static androidx.test.espresso.action.ViewActions.click;
 import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
+import static com.sdpteam.connectout.utils.FutureUtils.waitABit;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 
@@ -16,11 +17,8 @@ import org.junit.runner.RunWith;
 
 import com.sdpteam.connectout.R;
 import com.sdpteam.connectout.event.creator.EventCreatorActivity;
-import com.sdpteam.connectout.event.EventFirebaseDataSource;
-import com.sdpteam.connectout.event.nearbyEvents.EventsViewModel;
 import com.sdpteam.connectout.event.creator.LocationPicker;
 
-import android.os.SystemClock;
 import androidx.test.espresso.intent.Intents;
 import androidx.test.ext.junit.rules.ActivityScenarioRule;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
@@ -45,7 +43,7 @@ public class MapFragmentTest {
     public void testMapDisplayed() {
         onView(withId(R.id.map)).check(matches(isDisplayed()));
         onView(withId(R.id.refresh_button)).perform(click());
-        SystemClock.sleep(1000);
+        waitABit();
         onView(withId(R.id.map)).check(matches(isDisplayed()));
     }
