+++ app/src/androidTest/java/com/sdpteam/connectout/profile/ReportProfileActivityTest.java
@@ -8,31 +8,28 @@ import static androidx.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static com.sdpteam.connectout.profile.EditProfileActivity.NULL_USER;
 import static com.sdpteam.connectout.profile.ReportProfileActivity.REPORTED_UID;
-import static com.sdpteam.connectout.utils.FutureUtil.fJoin;
+import static com.sdpteam.connectout.utils.FutureUtils.fJoin;
+import static com.sdpteam.connectout.utils.FutureUtils.waitABit;
 import static org.junit.Assert.assertEquals;
-
-import android.content.Intent;
-import android.os.SystemClock;
-
-import androidx.test.core.app.ApplicationProvider;
-import androidx.test.espresso.Espresso;
-import androidx.test.ext.junit.rules.ActivityScenarioRule;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
-
-import com.sdpteam.connectout.R;
-import com.sdpteam.connectout.authentication.AuthenticatedUser;
-import com.sdpteam.connectout.authentication.GoogleAuth;
 
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
+import com.sdpteam.connectout.R;
+import com.sdpteam.connectout.authentication.AuthenticatedUser;
+import com.sdpteam.connectout.authentication.GoogleAuth;
+
+import android.content.Intent;
+import androidx.test.core.app.ApplicationProvider;
+import androidx.test.espresso.Espresso;
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
+
 @RunWith(AndroidJUnit4.class)
 public class ReportProfileActivityTest {
 
     public static String reportedUid = "testuid";
-    private final ReportFirebaseDataSource model = new ReportFirebaseDataSource();
-
     static Intent intent;
 
     static {
@@ -40,6 +37,7 @@ public class ReportProfileActivityTest {
         intent.putExtra(REPORTED_UID, reportedUid);
     }
 
+    private final ReportFirebaseDataSource model = new ReportFirebaseDataSource();
     @Rule
     public ActivityScenarioRule<ReportProfileActivity> testRule = new ActivityScenarioRule<>(intent);
 
@@ -56,10 +54,10 @@ public class ReportProfileActivityTest {
         onView(withId(R.id.ReportText)).perform(typeText("test report"));
         Espresso.closeSoftKeyboard();
         onView(withId(R.id.submitReportButton)).perform(click());
-        SystemClock.sleep(1000);
+        waitABit();
 
         AuthenticatedUser au = new GoogleAuth().loggedUser();
         String reporterUid = (au == null) ? NULL_USER : au.uid;
-        assertEquals("test report" , fJoin(model.fetchReport(reportedUid, reporterUid)));
+        assertEquals("test report", fJoin(model.fetchReport(reportedUid, reporterUid)));
     }
 }
