+++ app/src/androidTest/java/com/sdpteam/connectout/utils/FutureUtils.java
@@ -5,7 +5,9 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-public class FutureUtil {
+import android.os.SystemClock;
+
+public class FutureUtils {
     /**
      * join a future with timeout
      */
@@ -15,5 +17,12 @@ public class FutureUtil {
         } catch (ExecutionException | InterruptedException | TimeoutException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    /**
+     * Use this one for sleeps if you don't have access to the future
+     */
+    public static void waitABit() {
+        SystemClock.sleep(1500);
     }
 }
