+++ app/src/androidTest/java/com/sdpteam/connectout/validation/EventCreationValidatorTest.java
@@ -1,11 +1,13 @@
 package com.sdpteam.connectout.validation;
 
-import android.content.Context;
-import android.widget.EditText;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
 import org.junit.Test;
 
-import static org.junit.Assert.*;
-
+import android.content.Context;
+import android.widget.EditText;
 import androidx.test.platform.app.InstrumentationRegistry;
 
 public class EventCreationValidatorTest {
