+++ app/src/androidTest/java/com/sdpteam/connectout/validation/ValidationUtilsTest.java
@@ -1,11 +1,14 @@
 package com.sdpteam.connectout.validation;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
+
 import org.junit.Test;
 
 import android.content.Context;
 import android.widget.EditText;
-
 import androidx.test.platform.app.InstrumentationRegistry;
 
 public class ValidationUtilsTest {
